Rename to FastNoiseLite
