+++ PreviewApp/FastNoiseLite.csproj
@@ -8,6 +8,6 @@
     <PackageReference Include="Eto.Platform.Wpf" Version="2.5.2" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="..\CSharp\FastNoise.cs" Link="FastNoise.cs" />
+    <Compile Include="..\CSharp\FastNoiseLite.cs" Link="FastNoiseLite.cs" />
   </ItemGroup>
 </Project>
